* allow both 32 and 64 bit profilers to be registered at the same time (different GUIDs) * remedy outstanding 64-bit changes

+++ main/OpenCover.Profiler/Method.cpp
@@ -535,7 +535,7 @@ long Method::GetMethodSize()
     {
         long align = sizeof(DWORD) - 1;
         size = ((size + align) & ~align);
-        size += ((m_exceptions.size() * 6) + 1) * sizeof(long);
+        size += (((long)m_exceptions.size() * 6) + 1) * sizeof(long);
     }
 
     return size;
